requirements.txt
numpy
opencv-python
Pillow
gradio
requests
tqdm
